telebot/requirements.txt
telebot~=0.0.5
mysql-connector-python~=8.2.0
python-dotenv~=1.0.0
requests~=2.31.0